+++ src/lib/components/events/ChatBubble.svelte
@@ -1,15 +1,19 @@
-<div class="chat chat-end">
+<div class="flex justify-center">
+  <div
+    class="chat chat-end fixed z-20 bottom-0 left-0 pr-4 pb-4 md:p-0 backdrop-blur-lg md:backdrop-blur-0 md:relative w-full"
+  >
     <div class="chat-image avatar">
       <div class="w-10 rounded-full">
-        <img
-          alt="El avatar de Bovedin"
-          src="/bovedin.avif" />
+        <img alt="El avatar de Bovedin" src="/bovedin.avif" />
       </div>
     </div>
     <div class="chat-header">
       Bovedin
       <time class="text-xs opacity-50">12:46</time>
     </div>
-    <div class="chat-bubble chat-bubble-success">Tienes un código de descuento aplicado!</div>
+    <div class="chat-bubble chat-bubble-success md:text-nowrap md:max-w-fit">
+      Tienes un código de descuento aplicado!
+    </div>
     <!-- <div class="chat-footer opacity-50">Seen at 12:46</div> -->
   </div>
+</div>