+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Add
 
 - Add texture to some titles
+- Add discount codes in query params
 
 ## [0.0.42] - 2024-07-18